requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
altair
pyyaml
openpyxl
streamlit